requirements.txt
praw
requests
python-dotenv
textblob
vaderSentiment
